Allow only two ways to define a flow
