+++ test/micro/case/call_test.rb
@@ -10,14 +10,6 @@ class Micro::Case::CallTest < Minitest::Test
     Steps::Add2
   ])
 
-  class Add2ToAllNumbers2
-    include Micro::Case::Flow
-
-    flow Steps::ConvertToNumbers, Steps::Add2
-  end
-
-  Add2ToAllNumbers3 = Steps::ConvertToNumbers >> Steps::Add2
-
   def test_the_calling_of_use_cases
     assert_raises(ArgumentError) { Steps::ConvertToNumbers.new() }
 
@@ -31,19 +23,5 @@ class Micro::Case::CallTest < Minitest::Test
 
     assert_instance_of(Failure, Add2ToAllNumbers1.call({}))
     assert_instance_of(Failure, Add2ToAllNumbers1.call)
-  end
-
-  def test_the_calling_of_flow_classes
-    assert_raises(ArgumentError) { Add2ToAllNumbers2.new() }
-    assert_instance_of(Failure, Add2ToAllNumbers2.new({}).call)
-    assert_instance_of(Failure, Add2ToAllNumbers2.call({}))
-    assert_instance_of(Failure, Add2ToAllNumbers2.call)
-  end
-
-  def test_the_calling_of_a_flow_created_from_composition_operators
-    assert_raises(NoMethodError) { Add2ToAllNumbers3.new({}).call }
-
-    assert_instance_of(Failure, Add2ToAllNumbers3.call({}))
-    assert_instance_of(Failure, Add2ToAllNumbers3.call)
   end
 end
